+++ packages/dev-env/src/mock/index.ts
@@ -1,10 +1,6 @@
 import { AtUri } from '@atproto/syntax'
-import AtpAgent from '@atproto/api'
+import AtpAgent, { COM_ATPROTO_MODERATION } from '@atproto/api'
 import { Database } from '@atproto/bsky'
-import {
-  REASONSPAM,
-  REASONOTHER,
-} from '@atproto/api/src/client/types/com/atproto/moderation/defs'
 import { EXAMPLE_LABELER, TestNetwork } from '../index'
 import { postTexts, replyTexts } from './data'
 import labeledImgB64 from './img/labeled-img-b64'
@@ -15,13 +11,12 @@ import blurHashB64 from './img/blur-hash-avatar-b64'
 // we use this to ensure the mock dataset is always the same
 // which is very useful when testing
 // (not everything is currently deterministic but it could be)
-function* dateGen() {
+function* dateGen(): Generator<string, never> {
   let start = 1657846031914
   while (true) {
     yield new Date(start).toISOString()
     start += 1e3
   }
-  return ''
 }
 
 export async function generateMockSetup(env: TestNetwork) {
@@ -119,7 +114,10 @@ export async function generateMockSetup(env: TestNetwork) {
   // Report one user
   const reporter = picka(users)
   await reporter.agent.api.com.atproto.moderation.createReport({
-    reasonType: picka([REASONSPAM, REASONOTHER]),
+    reasonType: picka([
+      COM_ATPROTO_MODERATION.DefsReasonSpam,
+      COM_ATPROTO_MODERATION.DefsReasonOther,
+    ]),
     reason: picka(["Didn't look right to me", undefined, undefined]),
     subject: {
       $type: 'com.atproto.admin.defs#repoRef',
@@ -169,7 +167,10 @@ export async function generateMockSetup(env: TestNetwork) {
     if (rand(6) === 0) {
       const reporter = picka(users)
       await reporter.agent.api.com.atproto.moderation.createReport({
-        reasonType: picka([REASONSPAM, REASONOTHER]),
+        reasonType: picka([
+          COM_ATPROTO_MODERATION.DefsReasonSpam,
+          COM_ATPROTO_MODERATION.DefsReasonOther,
+        ]),
         reason: picka(["Didn't look right to me", undefined, undefined]),
         subject: {
           $type: 'com.atproto.repo.strongRef',
@@ -348,328 +349,6 @@ export async function generateMockSetup(env: TestNetwork) {
       createdAt: date.next().value,
     },
   )
-
-  // create the dev-env moderator
-  {
-    const res = await clients.loggedout.api.com.atproto.server.createAccount({
-      email: 'mod-authority@test.com',
-      handle: 'mod-authority.test',
-      password: 'hunter2',
-    })
-    const agent = env.pds.getClient()
-    agent.api.setHeader('Authorization', `Bearer ${res.data.accessJwt}`)
-    await agent.api.app.bsky.actor.profile.create(
-      { repo: res.data.did },
-      {
-        displayName: 'Dev-env Moderation',
-        description: `The pretend version of mod.bsky.app`,
-      },
-    )
-
-    await agent.api.app.bsky.labeler.service.create(
-      { repo: res.data.did, rkey: 'self' },
-      {
-        policies: {
-          labelValues: [
-            '!hide',
-            '!warn',
-            'porn',
-            'sexual',
-            'nudity',
-            'sexual-figurative',
-            'graphic-media',
-            'self-harm',
-            'sensitive',
-            'extremist',
-            'intolerant',
-            'threat',
-            'rude',
-            'illicit',
-            'security',
-            'unsafe-link',
-            'impersonation',
-            'misinformation',
-            'scam',
-            'engagement-farming',
-            'spam',
-            'rumor',
-            'misleading',
-            'inauthentic',
-          ],
-          labelValueDefinitions: [
-            {
-              identifier: 'spam',
-              blurs: 'content',
-              severity: 'inform',
-              defaultSetting: 'hide',
-              adultOnly: false,
-              locales: [
-                {
-                  lang: 'en',
-                  name: 'Spam',
-                  description:
-                    'Unwanted, repeated, or unrelated actions that bother users.',
-                },
-              ],
-            },
-            {
-              identifier: 'impersonation',
-              blurs: 'none',
-              severity: 'inform',
-              defaultSetting: 'hide',
-              adultOnly: false,
-              locales: [
-                {
-                  lang: 'en',
-                  name: 'Impersonation',
-                  description:
-                    'Pretending to be someone else without permission.',
-                },
-              ],
-            },
-            {
-              identifier: 'scam',
-              blurs: 'content',
-              severity: 'alert',
-              defaultSetting: 'hide',
-              adultOnly: false,
-              locales: [
-                {
-                  lang: 'en',
-                  name: 'Scam',
-                  description: 'Scams, phishing & fraud.',
-                },
-              ],
-            },
-            {
-              identifier: 'intolerant',
-              blurs: 'content',
-              severity: 'alert',
-              defaultSetting: 'warn',
-              adultOnly: false,
-              locales: [
-                {
-                  lang: 'en',
-                  name: 'Intolerance',
-                  description: 'Discrimination against protected groups.',
-                },
-              ],
-            },
-            {
-              identifier: 'self-harm',
-              blurs: 'content',
-              severity: 'alert',
-              defaultSetting: 'warn',
-              adultOnly: false,
-              locales: [
-                {
-                  lang: 'en',
-                  name: 'Self-Harm',
-                  description:
-                    'Promotes self-harm, including graphic images, glorifying discussions, or triggering stories.',
-                },
-              ],
-            },
-            {
-              identifier: 'security',
-              blurs: 'content',
-              severity: 'alert',
-              defaultSetting: 'hide',
-              adultOnly: false,
-              locales: [
-                {
-                  lang: 'en',
-                  name: 'Security Concerns',
-                  description:
-                    'May be unsafe and could harm your device, steal your info, or get your account hacked.',
-                },
-              ],
-            },
-            {
-              identifier: 'misleading',
-              blurs: 'content',
-              severity: 'alert',
-              defaultSetting: 'warn',
-              adultOnly: false,
-              locales: [
-                {
-                  lang: 'en',
-                  name: 'Misleading',
-                  description:
-                    'Altered images/videos, deceptive links, or false statements.',
-                },
-              ],
-            },
-            {
-              identifier: 'threat',
-              blurs: 'content',
-              severity: 'inform',
-              defaultSetting: 'hide',
-              adultOnly: false,
-              locales: [
-                {
-                  lang: 'en',
-                  name: 'Threats',
-                  description:
-                    'Promotes violence or harm towards others, including threats, incitement, or advocacy of harm.',
-                },
-              ],
-            },
-            {
-              identifier: 'unsafe-link',
-              blurs: 'content',
-              severity: 'alert',
-              defaultSetting: 'hide',
-              adultOnly: false,
-              locales: [
-                {
-                  lang: 'en',
-                  name: 'Unsafe link',
-                  description:
-                    'Links to harmful sites with malware, phishing, or violating content that risk security and privacy.',
-                },
-              ],
-            },
-            {
-              identifier: 'illicit',
-              blurs: 'content',
-              severity: 'alert',
-              defaultSetting: 'hide',
-              adultOnly: false,
-              locales: [
-                {
-                  lang: 'en',
-                  name: 'Illicit',
-                  description:
-                    'Promoting or selling potentially illicit goods, services, or activities.',
-                },
-              ],
-            },
-            {
-              identifier: 'misinformation',
-              blurs: 'content',
-              severity: 'inform',
-              defaultSetting: 'warn',
-              adultOnly: false,
-              locales: [
-                {
-                  lang: 'en',
-                  name: 'Misinformation',
-                  description:
-                    'Spreading false or misleading info, including unverified claims and harmful conspiracy theories.',
-                },
-              ],
-            },
-            {
-              identifier: 'rumor',
-              blurs: 'content',
-              severity: 'inform',
-              defaultSetting: 'warn',
-              adultOnly: false,
-              locales: [
-                {
-                  lang: 'en',
-                  name: 'Rumor',
-                  description:
-                    'Approach with caution, as these claims lack evidence from credible sources.',
-                },
-              ],
-            },
-            {
-              identifier: 'rude',
-              blurs: 'content',
-              severity: 'inform',
-              defaultSetting: 'hide',
-              adultOnly: false,
-              locales: [
-                {
-                  lang: 'en',
-                  name: 'Rude',
-                  description:
-                    'Rude or impolite, including crude language and disrespectful comments, without constructive purpose.',
-                },
-              ],
-            },
-            {
-              identifier: 'extremist',
-              blurs: 'content',
-              severity: 'alert',
-              defaultSetting: 'hide',
-              adultOnly: false,
-              locales: [
-                {
-                  lang: 'en',
-                  name: 'Extremist',
-                  description:
-                    'Radical views advocating violence, hate, or discrimination against individuals or groups.',
-                },
-              ],
-            },
-            {
-              identifier: 'sensitive',
-              blurs: 'content',
-              severity: 'alert',
-              defaultSetting: 'warn',
-              adultOnly: false,
-              locales: [
-                {
-                  lang: 'en',
-                  name: 'Sensitive',
-                  description:
-                    'May be upsetting, covering topics like substance abuse or mental health issues, cautioning sensitive viewers.',
-                },
-              ],
-            },
-            {
-              identifier: 'engagement-farming',
-              blurs: 'content',
-              severity: 'alert',
-              defaultSetting: 'hide',
-              adultOnly: false,
-              locales: [
-                {
-                  lang: 'en',
-                  name: 'Engagement Farming',
-                  description:
-                    'Insincere content or bulk actions aimed at gaining followers, including frequent follows, posts, and likes.',
-                },
-              ],
-            },
-            {
-              identifier: 'inauthentic',
-              blurs: 'content',
-              severity: 'alert',
-              defaultSetting: 'hide',
-              adultOnly: false,
-              locales: [
-                {
-                  lang: 'en',
-                  name: 'Inauthentic Account',
-                  description: 'Bot or a person pretending to be someone else.',
-                },
-              ],
-            },
-            {
-              identifier: 'sexual-figurative',
-              blurs: 'media',
-              severity: 'none',
-              defaultSetting: 'show',
-              adultOnly: true,
-              locales: [
-                {
-                  lang: 'en',
-                  name: 'Sexually Suggestive (Cartoon)',
-                  description:
-                    'Art with explicit or suggestive sexual themes, including provocative imagery or partial nudity.',
-                },
-              ],
-            },
-          ],
-        },
-        createdAt: date.next().value,
-      },
-    )
-  }
 
   // create a labeler account
   {